Switch version comparison cast from int to float.
fixes #3564
https://pulp.plan.io/issues/3564

(cherry picked from commit 3603338293bd6c0975f1661e09b15c556ba6c7ce)

+++ plugins/pulp_rpm/yum_plugin/util.py
@@ -205,12 +205,12 @@ def is_version_newer(new_version, old_version):
     if not new_version:
         return False
 
-    new = int(new_version)
+    new = float(new_version)
 
     if not old_version:
         old = 0
     else:
-        old = int(old_version)
+        old = float(old_version)
     return new > old
